feat(scripts): add `archlinux_util.sh` utility script for Arch Linux
This commit introduces a new utility script, `archlinux_util.sh`, under the `/scripts` directory. This script performs common checks for Arch-based systems and facilitates consistent usage of the pacman package manager. It also provides functions for adjusting directory permissions and installing packages.

This script is intended to be used as a utility helper when setting up and managing Arch Linux systems in devcontainer features.