fix(src/aws-cli): Remove option to select AWS CLI version, default to v1
This commit removes the option to select the AWS CLI version from the installation script. This change was made in response to AWS CLI v2 being moved to the Arch User Repository (AUR). Now, the script will always install AWS CLI v1 from the official repositories. A notice has been added to inform users of this change when the script is run.